--- magebench.txt
@@ -1,5 +1,0 @@
-make _gen.pbjs: 15, 16, 16, 15.550, 15.601 = 15.6302
-mage frontGen: 10, 11, 11, 10.549, 10.498 = 10.6094
-
-advantage: 5.0208
-time won: ~32%
